+++ sdk/eventhub/Azure.Messaging.EventHubs.Shared/tests/Infrastructure/IBlobEventLogger.cs
@@ -126,7 +126,7 @@ namespace Azure.Messaging.EventHubs.Tests
                                    string consumerGroup,
                                    string clientIdentifier,
                                    long sequenceNumber,
-                                   long replicationSegment,
+                                   int replicationSegment,
                                    long offset,
                                    string exception);
 
@@ -149,7 +149,7 @@ namespace Azure.Messaging.EventHubs.Tests
                                       string consumerGroup,
                                       string clientIdentifier,
                                       long sequenceNumber,
-                                      long replicationSegment,
+                                      int replicationSegment,
                                       long offset);
 
         /// <summary>
@@ -171,7 +171,7 @@ namespace Azure.Messaging.EventHubs.Tests
                                    string consumerGroup,
                                    string clientIdentifier,
                                    long sequenceNumber,
-                                   long replicationSegment,
+                                   int replicationSegment,
                                    long offset);
 
         /// <summary>
